Separate ConnectException from RequestException
