+++ phpstan-baseline.neon
@@ -211,12 +211,32 @@ parameters:
 			path: src/Exception/ConnectException.php
 
 		-
+			message: "#^Method GuzzleHttp\\\\Exception\\\\ConnectException\\:\\:setHandlerContext\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Exception/ConnectException.php
+
+		-
+			message: "#^Property GuzzleHttp\\\\Exception\\\\ConnectException\\:\\:\\$handlerContext type has no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Exception/ConnectException.php
+
+		-
+			message: "#^Method GuzzleHttp\\\\Exception\\\\ConnectException\\:\\:getHandlerContext\\(\\) return type has no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Exception/ConnectException.php
+
+		-
 			message: "#^Property GuzzleHttp\\\\Exception\\\\RequestException\\:\\:\\$handlerContext type has no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Exception/RequestException.php
 
 		-
 			message: "#^Method GuzzleHttp\\\\Exception\\\\RequestException\\:\\:__construct\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
+			count: 1
+			path: src/Exception/RequestException.php
+
+		-
+			message: "#^Method GuzzleHttp\\\\Exception\\\\RequestException\\:\\:setHandlerContext\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Exception/RequestException.php
