+++ tests/Exception/ConnectExceptionTest.php
@@ -4,6 +4,8 @@ namespace GuzzleHttp\Tests\Exception;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Psr7\Request;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Client\NetworkExceptionInterface;
+use Psr\Http\Client\RequestExceptionInterface;
 
 /**
  * @covers \GuzzleHttp\Exception\ConnectException
@@ -15,6 +17,8 @@ class ConnectExceptionTest extends TestCase
         $req = new Request('GET', '/');
         $prev = new \Exception();
         $e = new ConnectException('foo', $req, $prev, ['foo' => 'bar']);
+        self::assertInstanceOf(NetworkExceptionInterface::class, $e);
+        self::assertNotInstanceOf(RequestExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertNull($e->getResponse());
         self::assertFalse($e->hasResponse());
